requirements.txt
autopep8==1.5.7
Brotli==1.0.9
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cycler==0.10.0
dash==2.0.0
dash-bootstrap-components==0.13.0
dash-colorscales==0.0.4
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dtale==1.56.0
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Compress==1.10.1
flask-ngrok==0.0.25
future==0.18.2
google-api-core==2.0.1
google-auth==2.2.1
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.28.0
google-cloud-bigquery-storage==2.9.0
google-cloud-core==2.0.0
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
greenlet==1.1.1
grpcio==1.41.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kaleido==0.2.1
kiwisolver==1.3.2
libcst==0.3.21
lz4==3.1.3
MarkupSafe==2.0.1
matplotlib==3.4.3
missingno==0.4.2
mssql==1.0.1
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.2
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
pandas-gbq==0.15.0
patsy==0.5.1
Pillow==8.3.2
plotly==5.3.1
ppscore==1.2.0
proto-plus==1.19.2
protobuf==3.18.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydata-google-auth==1.2.0
pyodbc==4.0.32
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.24
squarify==0.4.3
statsmodels==0.12.2
strsimpy==0.2.1
tenacity==8.0.1
threadpoolctl==2.2.0
toml==0.10.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.1
xarray==0.19.0
xlrd==2.0.1
zipp==3.5.0
